@@ -1,16 +1,14 @@
 using System.Linq;
 
-namespace Tournament.Football
+namespace Tournament.Football;
+
+public class ChampionsLeagueResult
 {
-    public class ChampionsLeagueResult
-    {
 
-        public ChampionsLeagueResult(ChampionsLeague stage) => Stage = stage;
+    public ChampionsLeagueResult(ChampionsLeague stage) => Stage = stage;
 
-        public bool IsCompleted => Stage.Stages.Last().IsCompleted;
+    public bool IsCompleted => Stage.Stages.Last().IsCompleted;
 
-        public ChampionsLeague Stage { get; }
-
-    }
+    public ChampionsLeague Stage { get; }
 
 }
